+merge from testing
git-svn-id: https://svn.code.sf.net/p/remotefs/code/trunk@1142 7764f161-3e89-4957-9ae6-8e1f4830ae4a

+++ src/command.h
@@ -85,7 +85,7 @@ struct answer
 	uint32_t command;
 	uint32_t data_len;
 	int32_t ret;
-	int16_t ret_errno;
+	int32_t ret_errno;
 };
 
 /** get description of command. debug only */
